MachinLearningMicroservices/mds_video/requirements.txt
pandas
pywebhdfs
redis
rq
kafka-python